Factored out VirtualTableModule into a separate file
